+++ Src/Application/Object/UI/PlayerHp/PlayerHp.cpp
@@ -12,7 +12,7 @@ void PlayerHp::Update()
 	maxHp = PlayerStatus::Instance().GetValue("MAXHP");
 
 	// 切り取り範囲の変更
-	rectX = m_bar.pTex->GetWidth() * ((float)hp / maxHp);
+	rectX = (int)(m_bar.pTex->GetWidth() * ((float)hp / maxHp));
 	m_bar.rect.width = rectX;
 
 	// 座標の移動
@@ -22,10 +22,10 @@ void PlayerHp::Update()
 void PlayerHp::DrawSprite()
 {
 	// フレーム
-	KdShaderManager::Instance().m_spriteShader.DrawTex(m_frame.pTex, m_frame.pos.x, m_frame.pos.y, m_frame.pTex->GetWidth(), m_frame.pTex->GetHeight(), &m_frame.rect);
+	KdShaderManager::Instance().m_spriteShader.DrawTex(m_frame.pTex, (int)m_frame.pos.x, (int)m_frame.pos.y, m_frame.pTex->GetWidth(), m_frame.pTex->GetHeight(), &m_frame.rect);
 
 	// バー
-	KdShaderManager::Instance().m_spriteShader.DrawTex(m_bar.pTex, m_bar.pos.x, m_bar.pos.y, rectX, m_bar.pTex->GetHeight(), &m_bar.rect);
+	KdShaderManager::Instance().m_spriteShader.DrawTex(m_bar.pTex, (int)m_bar.pos.x, (int)m_bar.pos.y, rectX, m_bar.pTex->GetHeight(), &m_bar.rect);
 }
 
 void PlayerHp::Init()
